% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo.list}
\alias{geo.list}
\title{geo.list}
\format{
An object of class \code{list} of length 6.
}
\usage{
geo.list
}
\description{
List of sf objects that feature various geographic layers to map by.
}
\keyword{datasets}
